[glibutil] Allow specifying per-module logging function. JB#48496
It's rarely required. One example when it is useful is a module that
intercepts logging requests and forwards them somewhere - this way
it can avoid forwarding its own internal log to itself.

This eats up one reserved field in struct glog_module, keeping
the structure binary compatibility with the old code.