[glibutil] Initialize default log settings from the environment
The following environment variables are checked on Linux:

  GUTIL_LOG_DEFAULT_LEVEL - configures default log level
  GUTIL_LOG_TIMESTAMP     - configures gutil_log_timestamp option
  GUTIL_LOG_TID           - configures gutil_log_tid option

This setup is done once at program's startup and then the program
can override these settings the way it wants, based on its command
line options, configuration files or whatever.

Allows to add thread id prefix to the log messages. That can be useful
for debugging race conditions and other multi-threaded issues.