[glibutil] Explicitly ignore status in gutil_inotify_read. JB#61905
Unchecked return value identified by Coverity.

Out of the four possible statuses, G_IO_STATUS_AGAIN should not be
returned unless the channel is non-blocking and G_IO_STATUS_EOF should
not be returned at all in the inotify case. The inotify FD is poll'ed
internally, so there is no need for looping here to deal with the case
that more events are available for reading that do not fit the buffer.

So this looks like a false positive -> making it explicit.